Renamed internal/server to internal/httpserver
Addition of the protocol is a better package name than if it was not
included.
