+++ internal/httpserver/middleware.go
@@ -1,4 +1,4 @@
-package server
+package httpserver
 
 import (
 	"net/http"
